+++ bin/index.js
@@ -32,7 +32,8 @@ Local development:
     --env <environment>       => optionally sets the environment to display information for, defaults to 'dev'
     --cwd <path>              => optionally set the path to project, defaults to current working directory
   serve                       => start a local development server using local API files
-    --port <port>             => sets the port to use for server, defaults to 3000
+    --env <environment>       => optionally sets the environment to load environment variables from, defaults to 'dev'
+    --port <port>             => optionally sets the port to use for server, defaults to 3000
     --cwd <path>              => optionally set the path to project, defaults to current working directory
 
 Initial settings: